修正ソース取り込み

+++ kadai_002/samuraitravel/src/main/java/com/example/samuraitravel/repository/ReviewRepository.java
@@ -13,7 +13,7 @@ import com.example.samuraitravel.entity.User;
 public interface ReviewRepository extends JpaRepository<Review, Integer> {
 	 public Page<Review> findByHouseOrderByCreatedAtDesc(House review, Pageable pageable);
 	 public Page<Review> findByUserOrderByCreatedAtDesc(User review, Pageable pageable);
-	 public List<Review> findByHouseIdOrderByCreatedAtDesc(Long houseId);
+	 public List<Review> findByHouseIdOrderByCreatedAtDesc(int houseId);
